requirements.txt
# Core
pandas>=2.2
python-dotenv>=1.0

# LangChain + OpenAI
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-experimental>=0.0.58
openai>=1.55

# UI (needed for the Streamlit app; safe to keep for both)
streamlit>=1.41